requirements.txt
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
google-cloud-bigquery>=3.13.0
pandas>=2.0.0
python-dotenv>=1.0.0
db-dtypes>=1.2.0
pytest>=8.0.0
